fix sonarqube bug
